Rename file and class
- Docs の例に合わせてファイル名を変更(main.js → memo.js)
- 上記修正に伴い、memo.js → memo_model.js に修正
- ファイル名の修正に伴い、ファイル名に合わせる形でクラス名を修正(Memo → MemoModel)
