--- 04.class/main.js
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-import MemoApp from "./memo_app.js";
-
-new MemoApp().handleOption();
